+++ patzilla-ui/navigator/components/viewport.js
@@ -39,6 +39,9 @@ ViewportPlugin = Marionette.Controller.extend({
         }
     },
     document_rate: function(score, dismiss) {
+        if (navigatorApp.component_enabled('stack')) {
+            navigatorApp.stack.activate_rating_by_viewport();
+        }
         dismiss = dismiss || false;
         var document_number = this.get_document_number_in_focus();
         return this.app.document_rate(document_number, score, dismiss);