+++ patzilla-ui/navigator/style/index.js
@@ -1,2 +1,3 @@
 require('./generic.css');
+require('./elements.css');
 require('./app.css');